Rename headers class to MessageHeaders
